// frontend/src/app/core/api/api/models/customerDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomerDto { 
    customerId?: number;
    firstname?: string;
    lastname?: string;
    email?: string;
    phone?: string;
    birthDate?: string;
    drivingLicense?: string;
    streetName?: string;
    houseNumber?: string;
    postCode?: string;
    city?: string;
    country?: string;
    createdWhen?: string;
}